Попыталась сделать SpringBoot, надо еще доделать.
