+++ .idea/compiler.xml
@@ -7,13 +7,18 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="Payment" />
+        <module name="Server" />
+        <module name="Application" />
+        <module name="User" />
       </profile>
     </annotationProcessing>
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
-      <module name="Payment" options="-parameters" />
+      <module name="Application" options="-parameters" />
+      <module name="Payment" options="" />
+      <module name="Server" options="-parameters" />
+      <module name="User" options="-parameters" />
     </option>
   </component>
 </project>
