+++ User/src/main/java/ru/sbrf/User/UserAccountNumber.java
@@ -1,0 +1,20 @@
+package ru.sbrf.User;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class UserAccountNumber {
+    private String accountNumber;
+
+    public UserAccountNumber() {
+
+    }
+
+    public String getAccountNumber(String a) {
+        return a;
+    }
+}
+
+
