+++ User/src/main/java/ru/sbrf/User/UserNumber.java
@@ -1,0 +1,17 @@
+package ru.sbrf.User;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+
+@AllArgsConstructor
+@Getter
+public class UserNumber {
+    private String userNumber;
+
+    public UserNumber() {
+    }
+  }
+
+
+
