+++ User/src/main/java/ru/sbrf/User/Validation/ValidationUserNumber.java
@@ -1,0 +1,15 @@
+package ru.sbrf.User.Validation;
+
+import ru.sbrf.User.ExceptionUser;
+
+public class ValidationUserNumber {
+    private String userNumber;
+
+    public boolean validationUserNumber(String userNumber){
+        if (userNumber.length() != 8){
+            throw new ExceptionUser("invalid userNumber" + userNumber);
+        } else {
+            return true;
+        }
+    }
+}
